+++ mayan/apps/converter/apps.py
@@ -11,8 +11,14 @@ from mayan.apps.common.menus import (
     menu_list_facet, menu_multi_item, menu_object, menu_secondary,
     menu_setup
 )
+from mayan.apps.events.classes import EventModelRegistry, ModelEventType
+from mayan.apps.events.links import (
+    link_events_for_object, link_object_event_types_user_subcriptions_list,
+)
+from mayan.apps.events.permissions import permission_events_view
 from mayan.apps.navigation.classes import SourceColumn
 
+from .events import event_asset_edited
 from .links import (
     link_asset_create, link_asset_multiple_delete,
     link_asset_single_delete, link_asset_edit, link_asset_list,
@@ -38,11 +44,19 @@ class ConverterApp(MayanAppConfig):
         Asset = self.get_model(model_name='Asset')
         LayerTransformation = self.get_model(model_name='LayerTransformation')
 
+        EventModelRegistry.register(model=Asset)
+
+        ModelEventType.register(
+            model=Asset, event_types=(
+                event_asset_edited,
+            )
+        )
+
         ModelPermission.register(
             model=Asset, permissions=(
                 permission_acl_edit, permission_acl_view,
                 permission_asset_delete, permission_asset_edit,
-                permission_asset_view
+                permission_asset_view, permission_events_view
             )
         )
 
@@ -75,9 +89,11 @@ class ConverterApp(MayanAppConfig):
 
         menu_list_facet.bind_links(
             links=(
-                link_acl_list,
+                link_acl_list, link_events_for_object,
+                link_object_event_types_user_subcriptions_list,
             ), sources=(Asset,)
         )
+
         menu_multi_item.bind_links(
             links=(link_asset_multiple_delete,), sources=(Asset,)
         )