+++ mayan/apps/converter/tests/mixins.py
@@ -28,14 +28,22 @@ class AssetTestMixin:
 
 class AssetViewTestMixin:
     def _request_test_asset_create_view(self):
+        pk_list = list(Asset.objects.values_list('pk', flat=True))
+
         with open(file=TEST_ASSET_PATH, mode='rb') as file_object:
-            return self.post(
+            response = self.post(
                 viewname='converter:asset_create', data={
                     'label': TEST_ASSET_LABEL,
                     'internal_name': TEST_ASSET_INTERNAL_NAME,
                     'file': file_object
                 }
             )
+
+        self.test_asset = Asset.objects.exclude(
+            pk__in=pk_list
+        ).first()
+
+        return response
 
     def _request_test_asset_delete_view(self):
         return self.post(