+++ HISTORY.rst
@@ -220,6 +220,7 @@
   trigger.
 - Add the setting ``SEARCH_RESULTS_LIMIT`` which limits the number search
   results to fetch and display. Defaults to ``100``.
+- Add events to assets.
 
 3.4.15 (2020-08-26)
 ===================